+++ packages/lib/link/index.js
@@ -1,12 +1,15 @@
-/* eslint-disable global-require */
 /**
  * Library for message based communication between nodes
  * @module lib/link
  * @author Hornwitser
  */
 "use strict";
+const link = require("./link");
+const messages = require("./messages");
+const connectors = require("./connectors");
+
 module.exports = {
-	...require("./link"),
-	...require("./messages"),
-	...require("./connectors"),
+	...link,
+	...messages,
+	...connectors,
 };