+++ packages/lib/shared_commands.js
@@ -55,7 +55,7 @@ async function handlePluginCommand(args, pluginList, pluginListPath) {
 
 		let pluginInfo;
 		try {
-			// eslint-disable-next-line global-require
+			// eslint-disable-next-line node/global-require
 			pluginInfo = require(path.posix.join(pluginPath, "info"));
 		} catch (err) {
 			logger.error(`Unable to import plugin info from ${args.path}:\n${err.stack}`);